moveonly: add mp/type-threadmap.h
