+++ CMakeLists.txt
@@ -60,6 +60,7 @@ set(MP_PUBLIC_HEADERS
   include/mp/type-set.h
   include/mp/type-string.h
   include/mp/type-struct.h
+  include/mp/type-threadmap.h
   include/mp/type-tuple.h
   include/mp/type-vector.h
   include/mp/type-void.h
