+++ include/mp/proxy-types.h
@@ -297,35 +297,6 @@ void PassField(Priority<0>, TypeList<>, ServerContext& server_context, const Fn&
     BuildField(TypeList<>(), server_context, Make<StructField, Accessor>(results));
 }
 
-template <>
-struct ProxyServer<ThreadMap> final : public virtual ThreadMap::Server
-{
-public:
-    ProxyServer(Connection& connection);
-    kj::Promise<void> makeThread(MakeThreadContext context) override;
-    Connection& m_connection;
-};
-
-template <typename Output>
-void CustomBuildField(TypeList<>,
-    Priority<1>,
-    InvokeContext& invoke_context,
-    Output&& output,
-    typename std::enable_if<std::is_same<decltype(output.get()), ThreadMap::Client>::value>::type* enable = nullptr)
-{
-    output.set(kj::heap<ProxyServer<ThreadMap>>(invoke_context.connection));
-}
-
-template <typename Input>
-decltype(auto) CustomReadField(TypeList<>,
-    Priority<1>,
-    InvokeContext& invoke_context,
-    Input&& input,
-    typename std::enable_if<std::is_same<decltype(input.get()), ThreadMap::Client>::value>::type* enable = nullptr)
-{
-    invoke_context.connection.m_thread_map = input.get();
-}
-
 template <typename Derived, size_t N = 0>
 struct IterateFieldsHelper
 {
